Start adding bottle
